fix: place tests beside their source and tighten rename and placement
